Fix build: add VotingPlugin stub jar and fix compilation errors
- Add VotingPlugin.jar stub with VotingPluginUser, VoteSite, and
  VotingPluginMain classes needed for compilation
- Fix lossy double-to-int conversion in VotePluginIntegration.java
- Update .gitignore to track the stub jar

+++ runbad-bot-bridge/src/main/java/com/runbad/bridge/integrations/VotePluginIntegration.java
@@ -65,7 +65,7 @@ public class VotePluginIntegration {
 
                 // Get the time of last vote for this site
                 long lastVoteTime = vpUser.getTime(site);
-                int cooldownHours = site.getVoteDelay(); // cooldown in hours
+                int cooldownHours = (int) site.getVoteDelay(); // cooldown in hours
 
                 long cooldownMs = cooldownHours * 3600L * 1000L;
                 long nextVoteMs = lastVoteTime + cooldownMs;
